ref(exasol): Move exasol code out of contrib
